+++ DalaMock.Host/Hosting/HostedMessages.cs
@@ -1,0 +1,7 @@
+namespace DalaMock.Host.Hosting;
+
+using DalaMock.Host.Mediator;
+
+public record PluginStartedMessage() : MessageBase;
+
+public record PluginStoppingMessage() : MessageBase();