+++ DalaMock.Host/Hosting/HostedPluginOptions.cs
@@ -1,0 +1,9 @@
+namespace DalaMock.Host.Hosting;
+
+/// <summary>
+/// Configure the default services registered with your plugin.
+/// </summary>
+public class HostedPluginOptions
+{
+    public bool UseMediatorService { get; set; } = false;
+}