+++ DalaMock.Sample/Mock/DalamudMockPluginTest.cs
@@ -1,28 +1,22 @@
-﻿// <copyright file="DalamudMockPluginTest.cs" company="PlaceholderCompany">
-// Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-
-using Microsoft.Extensions.Logging;
-
-namespace DalaMock.Sample;
+﻿namespace DalaMock.Sample.Mock;
 
 using Autofac;
 using DalaMock.Core.Mocks;
 using DalaMock.Core.Windows;
 using DalaMock.Shared.Interfaces;
-using Dalamud.Interface.Windowing;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
+using Microsoft.Extensions.Logging;
 
 public class DalamudMockPluginTest : DalamudPluginTest
 {
     private readonly ILogger<DalamudMockPluginTest> logger;
 
-    public DalamudMockPluginTest(IDalamudPluginInterface pluginInterface, ILogger<DalamudMockPluginTest> logger,  IPluginLog pluginLog, IDataManager dataManager, ITextureProvider textureProvider, IChatGui chatGui, IDtrBar dtrBar)
-        : base(pluginInterface, pluginLog, dataManager, textureProvider, chatGui, dtrBar)
+    public DalamudMockPluginTest(IDalamudPluginInterface pluginInterface, ILogger<DalamudMockPluginTest> logger, IPluginLog pluginLog, IFramework framework, ICommandManager commandManager, IDataManager dataManager, ITextureProvider textureProvider, IChatGui chatGui, IDtrBar dtrBar)
+        : base(pluginInterface, pluginLog, framework, commandManager, dataManager, textureProvider, chatGui, dtrBar)
     {
         this.logger = logger;
-        this.logger.LogInformation("Plugin test started");
+        this.logger.LogInformation("Plugin mock started");
     }
 
     public override void ConfigureContainer(ContainerBuilder containerBuilder)