+++ DalaMock.Sample/TestMockService.cs
@@ -1,8 +1,0 @@
-﻿namespace DalaMock.Sample;
-
-using Core.Mocks;
-
-public class TestMockService : IMockService
-{
-    public string ServiceName { get; }
-}