+++ DalaMock.Sample/Windows/ConfigWindow.cs
@@ -1,0 +1,17 @@
+namespace DalaMock.Sample.Windows;
+
+using Dalamud.Interface.Windowing;
+using ImGuiNET;
+
+public class ConfigWindow(Configuration config) : Window("SamplePlugin Config")
+{
+    public override void Draw()
+    {
+        var configOption = config.ConfigOption;
+
+        if (ImGui.Checkbox("Config Option", ref configOption))
+        {
+            config.ConfigOption = configOption;
+        }
+    }
+}